// src/store/types/global.types.ts
export  enum GlobalActionTypes {
    SET_COORDINATES_OF_NODES = 'SET_COORDINATES_OF_NODES',
    SET_NUMBER_OF_CLUSTERS = 'SET_NUMBER_OF_CLUSTERS',
    SET_ALGORITHM = 'SET_ALGORITHM',
    RESET = 'RESET',
    START_VISUALIZATION = 'START_VISUALIZATION',
    END_VISUALIZATION = 'END_VISUALIZATION',
    SET_SPEED = 'SET_SPEED',
    UPDATE_COORDINATES = 'UPDATE_COORDINATES',
    SET_LEARN_MODE = 'SET_LEARN_MODE',
    SET_ITERATION_MODE='SET_ITERATION_MODE',
    SET_MAX_ITERATIONS ='SET_MAX_ITERATIONS'
}


export default GlobalActionTypes;